Reorganize JS into dedicated legend.js file
